+++ .vscode/settings.json
@@ -1,6 +1,9 @@
 {
     "files.associations": {
         "conio.h": "c",
-        "bros.h": "c"
+        "bros.h": "c",
+        "brosgui.h": "c",
+        "testapp1.h": "c",
+        "broskernel.h": "c"
     }
 }
