--- inc/bros2.h
@@ -1,8 +1,0 @@
-#ifndef __BROS_2_H__
-#define __BROS_2_H__
-
-// Just here to test the build system
-
-int test1();
-
-#endif
