+++ inc/testApp1.h
@@ -1,0 +1,8 @@
+#ifndef TEST_APP_1_H
+#define TEST_APP_1_H
+
+void testApp1Start();
+void testApp1Proc();
+void testApp1Exit();
+
+#endif
