+++ inc/testApp2.h
@@ -1,0 +1,8 @@
+#ifndef TEST_APP_2_H
+#define TEST_APP_2_H
+
+void testApp2Start();
+void testApp2Proc();
+void testApp2Exit();
+
+#endif
