--- src/bros2.c
@@ -1,6 +1,0 @@
-#include "bros2.h"
-
-// Nothing special. Here to test the build system.
-int test1() {
-    return 12;
-}
